requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
requests==2.31.0
gunicorn==21.2.0


